// src/app/modules/graficos/models/agenteCategory.ts
export interface AgenteCategory{
    businessName?: string;
    idAgente?: string;
    customizacao?: number;
    treinamentoOnline?: number;
    implantacao?: number;
    licitacao?: number;
    solicitacaoServico?: number;
    homologacao?: number;
    falha?: number;
    semCategoria?: number;
    solicitacaoTreinamento?: number;
    sugestao?: number;
    duvida?: number;
}